Remove the AdamSchafferFunctionN2Test test from the Adam test suite.
The AdamSchafferFunctionN2Test fails on some specific aarch64 machines, since we are unable to reproduce the issue on other devices, we decided to remove the test from the Adam test suite. The Adam optimizer is still tested against many different functions, which all pass, so we don't think the implementation is wrong. It might be possible to optimize the initial function and optimizer settings to help Adam to converge in certain situations, but since we can't reproduce the issue, it's unclear what those settings are. For more details, see #253.